+++ src/main/kotlin/app/revanced/patcher/extensions/Extensions.kt
@@ -1,7 +1,6 @@
 package app.revanced.patcher.extensions
 
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod
-import com.android.tools.smali.dexlib2.AccessFlags
 
 /**
  * Create a label for the instruction at given index.
@@ -10,24 +9,3 @@ import com.android.tools.smali.dexlib2.AccessFlags
  * @return The label.
  */
 fun MutableMethod.newLabel(index: Int) = implementation!!.newLabelForIndex(index)
-
-/**
- * Perform a bitwise OR operation between an [AccessFlags] and an [Int].
- *
- * @param other The [Int] to perform the operation with.
- */
-infix fun Int.or(other: AccessFlags) = this or other.value
-
-/**
- * Perform a bitwise OR operation between two [AccessFlags].
- *
- * @param other The other [AccessFlags] to perform the operation with.
- */
-infix fun AccessFlags.or(other: AccessFlags) = value or other.value
-
-/**
- * Perform a bitwise OR operation between an [Int] and an [AccessFlags].
- *
- * @param other The [AccessFlags] to perform the operation with.
- */
-infix fun AccessFlags.or(other: Int) = value or other