+++ src/main/kotlin/app/revanced/patcher/patch/PatchResult.kt
@@ -1,9 +1,0 @@
-package app.revanced.patcher.patch
-
-/**
- * A result of executing a [Patch].
- *
- * @param patch The [Patch] that was executed.
- * @param exception The [PatchException] thrown, if any.
- */
-class PatchResult internal constructor(val patch: Patch<*>, val exception: PatchException? = null)