+++ src/main/kotlin/app/revanced/patcher/util/proxy/mutableTypes/encodedValue/MutableNullEncodedValue.kt
@@ -3,10 +3,10 @@ package app.revanced.patcher.util.proxy.mutableTypes.encodedValue
 import com.android.tools.smali.dexlib2.base.value.BaseNullEncodedValue
 import com.android.tools.smali.dexlib2.iface.value.ByteEncodedValue
 
-class MutableNullEncodedValue : BaseNullEncodedValue(), MutableEncodedValue {
+class MutableNullEncodedValue :
+    BaseNullEncodedValue(),
+    MutableEncodedValue {
     companion object {
-        fun ByteEncodedValue.toMutable(): MutableByteEncodedValue {
-            return MutableByteEncodedValue(this)
-        }
+        fun ByteEncodedValue.toMutable(): MutableByteEncodedValue = MutableByteEncodedValue(this)
     }
 }